+++ yash-semantics/src/command/pipeline.rs
@@ -36,7 +36,6 @@ use yash_env::stack::Frame;
 use yash_env::subshell::JobControl;
 use yash_env::subshell::Subshell;
 use yash_env::system::Errno;
-use yash_env::system::SystemEx;
 use yash_syntax::syntax;
 
 /// Executes the pipeline.
@@ -80,17 +79,6 @@ impl Command for syntax::Pipeline {
     async fn execute(&self, env: &mut Env) -> Result {
         if env.options.get(Exec) == Off && env.options.get(Interactive) == Off {
             return Continue(());
-        }
-
-        if env.controls_jobs() {
-            if let Ok(tty) = env.get_tty() {
-                // If we try to do job control when we're in the background,
-                // we'll mess up the terminal state controlled by the foreground
-                // shell. Calling tcsetpgrp_without_block from the background
-                // suspends our shell process until it's resumed in the
-                // foreground.
-                env.system.tcsetpgrp_without_block(tty, env.main_pgid).ok();
-            }
         }
 
         if !self.negation {